Move all cloze-related code to cloze.go
